backend/requirements.txt
fastapi
uvicorn[standard]
requests
mysql-connector-python
python-dotenv
groq
passlib[bcrypt]
pytest
pydantic[email]
python-jose